All the methods and enpoints related to the level of the groups were added

+++ src/main/java/ntnu/idatt2016/v233/SmartMat/entity/group/Group.java
@@ -34,10 +34,10 @@ public class Group {
     long groupId;
 
     @Column(name = "level")
-    int level;
+    long level;
 
     @Column(name = "points")
-    int points;
+    long points;
     
     @Column(name = "group_name")
     String groupName;
